requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
huggingface-hub==0.20.3
python-dotenv==1.0.1
